fix: Replace serialize_circuit function with `from<&Circuit>`
