+++ aztec_backend_wasm/src/lib.rs
@@ -2,7 +2,7 @@
 #![warn(unreachable_pub)]
 
 use common::acvm::acir::circuit::Circuit;
-use common::serializer::serialize_circuit;
+use common::barretenberg_structures::ConstraintSystem;
 use wasm_bindgen::prelude::*;
 
 #[wasm_bindgen]
@@ -10,5 +10,5 @@ pub fn serialize_acir_to_barretenberg_circuit(acir_bytes: Vec<u8>) -> Vec<u8> {
     console_error_panic_hook::set_once();
 
     let circuit = Circuit::read(&*acir_bytes).unwrap();
-    serialize_circuit(&circuit).to_bytes()
+    ConstraintSystem::from(&circuit).to_bytes()
 }
