+++ barretenberg_static_lib/src/acvm_interop/proof_system.rs
@@ -5,7 +5,6 @@ use common::acvm::FieldElement;
 use common::acvm::{Language, ProofSystemCompiler};
 use common::barretenberg_structures::Assignments;
 use common::proof;
-use common::serializer::serialize_circuit;
 use std::collections::BTreeMap;
 
 impl ProofSystemCompiler for Plonk {
@@ -14,9 +13,7 @@ impl ProofSystemCompiler for Plonk {
     }
 
     fn get_exact_circuit_size(&self, circuit: &Circuit) -> u32 {
-        let constraint_system = serialize_circuit(circuit);
-
-        StandardComposer::get_exact_circuit_size(&constraint_system)
+        StandardComposer::get_exact_circuit_size(&circuit.into())
     }
 
     fn black_box_function_supported(&self, opcode: &common::acvm::acir::BlackBoxFunc) -> bool {
@@ -38,8 +35,7 @@ impl ProofSystemCompiler for Plonk {
     }
 
     fn preprocess(&self, circuit: &Circuit) -> (Vec<u8>, Vec<u8>) {
-        let constraint_system = serialize_circuit(circuit);
-        let composer = StandardComposer::new(constraint_system);
+        let composer = StandardComposer::new(circuit.into());
 
         let proving_key = composer.compute_proving_key();
         let verification_key = composer.compute_verification_key(&proving_key);
@@ -53,8 +49,7 @@ impl ProofSystemCompiler for Plonk {
         witness_values: BTreeMap<Witness, FieldElement>,
         proving_key: &[u8],
     ) -> Vec<u8> {
-        let constraint_system = serialize_circuit(circuit);
-        let mut composer = StandardComposer::new(constraint_system);
+        let mut composer = StandardComposer::new(circuit.into());
 
         let assignments = proof::flatten_witness_map(circuit, witness_values);
 
@@ -68,8 +63,7 @@ impl ProofSystemCompiler for Plonk {
         circuit: &Circuit,
         verification_key: &[u8],
     ) -> bool {
-        let constraint_system = serialize_circuit(circuit);
-        let mut composer = StandardComposer::new(constraint_system);
+        let mut composer = StandardComposer::new(circuit.into());
 
         // Unlike when proving, we omit any unassigned witnesses.
         // Witness values should be ordered by their index but we skip over any indices without an assignment.
